Allow to override current time
from SOURCE_DATE_EPOCH environment variable
to allow for reproducible builds of nedit packages

See https://reproducible-builds.org/ for why this is good
and https://reproducible-builds.org/specs/source-date-epoch/
for the definition of this variable.

The date call is designed to work with both GNU date and BSD date.

Changes date format to ISO standard (the only truly universal format)